integrator updates.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@4820 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/math/distributed/integrator/integrator.hpp
@@ -32,7 +32,7 @@ struct integrator
         std::vector<naming::id_type> const& discovery_network
       , actions::function<T(T const&)> const& f
       , T const& tolerance
-      , boost::uint64_t regrid_segs 
+      , boost::uint32_t regrid_segs 
       , T const& eps 
     ) {
         return this->base_type::build_network_async
@@ -43,7 +43,7 @@ struct integrator
         std::vector<naming::id_type> const& discovery_network
       , actions::function<T(T const&)> const& f
       , T const& tolerance
-      , boost::uint64_t regrid_segs 
+      , boost::uint32_t regrid_segs 
       , T const& eps 
     ) {
         return this->base_type::build_network_sync
@@ -54,7 +54,7 @@ struct integrator
         std::vector<naming::id_type> const& discovery_network
       , actions::function<T(T const&)> const& f
       , T const& tolerance
-      , boost::uint64_t regrid_segs 
+      , boost::uint32_t regrid_segs 
       , T const& eps 
     ) {
         return this->base_type::build_network
@@ -65,7 +65,7 @@ struct integrator
     lcos::future_value<T> solve_async(
         T const& lower_bound
       , T const& upper_bound
-      , boost::uint64_t segments
+      , boost::uint32_t segments
     ) {
         return this->base_type::solve_async
             (this->gid_, lower_bound, upper_bound, segments);
@@ -74,7 +74,7 @@ struct integrator
     T solve_sync(
         T const& lower_bound
       , T const& upper_bound
-      , boost::uint64_t segments
+      , boost::uint32_t segments
     ) {
         return this->base_type::solve_sync
             (this->gid_, lower_bound, upper_bound, segments);
@@ -83,7 +83,7 @@ struct integrator
     T solve(
         T const& lower_bound
       , T const& upper_bound
-      , boost::uint64_t segments
+      , boost::uint32_t segments
     ) {
         return this->base_type::solve
             (this->gid_, lower_bound, upper_bound, segments);
